requirements.txt
# ---------------------------------------------------------------
# Python dependencies
# Install: pip install -r requirements.txt
# ---------------------------------------------------------------

# --- dbt core + DuckDB adapter ---
dbt-core==1.8.3
dbt-duckdb==1.8.2

# --- DuckDB standalone (for ad-hoc queries / scripts) ---
duckdb==1.0.0

# --- Apache Airflow (pin to match docker-compose image) ---
# Install only for local non-Docker Airflow usage.
# For Docker, the image handles its own dependencies.
apache-airflow==2.9.1

# --- Airflow providers ---
apache-airflow-providers-common-sql==1.14.0

# --- PySpark (for spark/process_historical.py) ---
pyspark==3.5.1

# --- Data utilities ---
pandas==2.2.2
pyarrow==16.1.0        # Parquet read/write support
fsspec==2024.5.0        # Filesystem abstractions

# --- Development / testing ---
pytest==8.2.1
pytest-dbt-core==0.1.0  # dbt test runner integration
black==24.4.2
ruff==0.4.7
pre-commit==3.7.1
